+++ packages/fpdart/lib/src/typedef.dart
@@ -2,4 +2,3 @@ import 'typeclass/hkt.dart';
 
 typedef Endo<A> = A Function(A a);
 typedef Separated<KT, A, B> = (HKT<KT, A>, HKT<KT, B>);
-typedef Magma<T> = T Function(T x, T y);